Vistas y foto de usuario
Las vistas en el nav-item de estadisticas, foto de usuario, vista del logueo y los nuevos js de ajax

+++ Museo/Models/VisitFlow/TypeAct.cs
@@ -12,8 +12,8 @@ namespace Museo.Models
         public string Description { get; set; }
 
         [Required]
-        [Range(0, int.MaxValue, ErrorMessage = "El costo debe ser mayor que cero")]
-        public int Cost { get; set; }
+        [Range(0, float.MaxValue, ErrorMessage = "El costo debe ser mayor que cero")]
+        public float Cost { get; set; }
 
         public string Name { get; set; }
     }
